Add art to the Cast receiver app
Change the saved art URI to be the remove URI as opposed to a
locally cached resource URI.
